Fix deprecation of `if_delegate_has_method`
The method is deprecated and was removed in sklearn 1.3.
To deal with this issue this commit introduces a fallback
upon `ImportError` and uses the recommended `if_available`
instead.

In the future once sklearn <1.0 is deprecated, we can use
`if_available` directly and remove the `if_delegate_has_method`
calls completely.